1.0.2: Invite users register route
A new route was added to support registering users via the invite
system. This route acts very similarly to the original register route
except that it will validate the token and invite email encoded in the
url sent to the invitee. No further confirmation is done as well.

The original registration route was also updated to support the
inviteOnly configuration boolean. When this boolean is set to true, the
public registration route is disabled. 1.0.2: Invite users register
route

A new route was added to support registering users via the invite
system. This route acts very similarly to the original register route
except that it will validate the token and invite email encoded in the
url sent to the invitee. No further confirmation is done as well.

The original registration route was also updated to support the
inviteOnly configuration boolean. When this boolean is set to true, the
public registration route is disabled.